Fixed a bug in Hidden pair strategy
For this example row : { 34789, 16, 3478, 234678, 3478, 3478,234689, 5, 348 },
Previous SolveForHiddenPairInRow method would accept 9 as a hidden pair
even tho it is only a single digit occurs twice whereas a hidden pair
should be a pair of digits both of which occuring only twice.

                if (hiddenPairDict.Count == 2)
fixed this problem using this check. Now only a pair of digits ocurring
twice are accepted as a hidden pair.

Before I forget, here is a nother small bug that's been fixed with this
commit:

GetDigitOccurrencesDictionaryInRow when trying to count zeros would throw:
System.Collections.Generic.KeyNotFoundException: The given key '0' was not present in the dictionary.

Because of the implementation the digit 0 is not included in the
dictionary. With the check if (cell == 0) continue; this expection is
now eliminated.